% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_hunter_formula_utils.R
\name{.build_linear_formula}
\alias{.build_linear_formula}
\title{Internal formula utilities for FeatureHunter}
\usage{
.build_linear_formula(intercept, coefs, names_vec, prefix = "logit(p) = ")
}
\description{
Internal formula utilities for FeatureHunter
}
\keyword{internal}
