% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_hunter_train.R
\name{.train_single_model}
\alias{.train_single_model}
\title{Internal training utilities for FeatureHunter}
\usage{
.train_single_model(
  model_type,
  train_exp,
  train_y,
  seed,
  fit_params,
  run_id = 1
)
}
\description{
Internal training utilities for FeatureHunter
}
\keyword{internal}
