% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_hunter_model_utils.R
\name{.uses_shap_for_type}
\alias{.uses_shap_for_type}
\title{Internal model utilities for FeatureHunter}
\usage{
.uses_shap_for_type(model_type)
}
\description{
Internal model utilities for FeatureHunter
}
\keyword{internal}
