% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fh_example_path.R
\name{fh_example_path}
\alias{fh_example_path}
\title{Path to example data shipped with the package}
\usage{
fh_example_path(...)
}
\arguments{
\item{...}{Character components appended under inst/extdata}
}
\value{
A file path (character scalar). Returns "" if not found.
}
\description{
Path to example data shipped with the package
}
\examples{
fh_example_path("raw_data_pncs")
fh_example_path("raw_data_pncs", "DatasetA.txt")
}
