% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{fh_init_tf}
\alias{fh_init_tf}
\title{Initialize TensorFlow / Keras Environment}
\usage{
fh_init_tf()
}
\value{
No return value.
Side effect: TensorFlow is initialized and set to Eager Execution mode if available.
}
\description{
Checks whether TensorFlow is available in the current Python environment.
If not installed, a message is displayed.
If installed, Eager Execution is enabled.
}
\examples{
\dontrun{
fh_init_tf()
}

}
