% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_logged_like}
\alias{is_logged_like}
\title{Check if data looks "log-scaled"}
\usage{
is_logged_like(
  mat,
  q_hi_prob = 0.99,
  q_hi_threshold = 20,
  frac_nonint_threshold = 0.2
)
}
\arguments{
\item{mat}{A matrix or data frame containing numeric values to be tested.}

\item{q_hi_prob}{(0.99)
99\%
Quantile probability (default 0.99). The high quantile (e.g., 99th percentile)
is used to approximate the upper bound of the data.}

\item{q_hi_threshold}{(20)
log
Threshold for the high quantile (default 20). If the computed quantile
does not exceed this threshold, the data are considered consistent with a
log-like scale.}

\item{frac_nonint_threshold}{(0.2)
log
Fraction of non-integer values required (default 0.2). A higher fraction
indicates that the data are less likely to be raw counts and more likely
to be log-transformed.}
}
\value{
TRUE FALSE"log"
A logical value, TRUE if the data resemble a log-like scale, FALSE otherwise.
}
\description{
log
This function heuristically determines whether the input values resemble
log-transformed data. Log-transformed values are usually within a limited
numeric range (often < 20) and contain a substantial proportion of
non-integer values.
}
