% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minority_prop}
\alias{minority_prop}
\title{Compute minority class proportion}
\usage{
minority_prop(y01)
}
\arguments{
\item{y01}{0/1 1
A binary label vector, ideally encoded as 0/1. Logical or numeric values are coerced to integer and compared to 1.}
}
\value{
\code{[0, 0.5]}
A numeric scalar in \code{[0, 0.5]}, giving the proportion of the minority class.
}
\description{
minority class
For a binary label vector, computes the proportion of the minority class.
}
\details{
1 \eqn{p1}
\eqn{\min(p1, 1 - p1)}
0.5 0
This function is often used as a helper for imbalance detection
(e.g. in \code{need_imbalance_handling()}).
}
\examples{
y <- c(0,0,0,1,1)
minority_prop(y) # 0.4
y2 <- c(1,1,1,1,0)
minority_prop(y2) # 0.2
}
\seealso{
\code{\link{need_imbalance_handling}}, \code{\link{class_weights_balanced}}
}
