% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_matrix}
\alias{safe_matrix}
\title{Safely coerce an object to a matrix}
\usage{
safe_matrix(x)
}
\arguments{
\item{x}{Input object: matrix, data.frame, or vector.}
}
\value{
A matrix object.
}
\description{
\code{as.matrix()}
Safely converts input to a matrix:
If already a matrix, return as-is;
If a data.frame, convert via \code{as.matrix()};
If a vector, wrap into a single-column matrix;
Otherwise, throw an error.
}
\examples{
safe_matrix(matrix(1:4, 2))
safe_matrix(data.frame(a=1:3, b=4:6))
safe_matrix(1:5)
\dontrun{
safe_matrix(list(a=1)) # 
}
}
